outbound-odespa/requirements.txt
fastapi
uvicorn
python-dotenv
plivo
pydantic
elevenlabs
supabase
sqlmodel
psycopg2